Major Update
This update includes a more domain focused structure, i.e. rather than having a package dedicated to resources/controllers, it's now focusing more on the specific function, i.e. the controller, service & so on for the product part of the application is bundled in the 'product' package. There has also been some validation implemented in such a way where I've tried to implement a fluent-interface type validation solution. Of course it's not so flexible that it could work for all potential solutions, but it's getting there.