Major Update
This commit includes a lot of changes around class names, it also includes some code related to spring security & the OWASP dependency scanner. Not to mention it includes some changes to the pom file such as jacoco tweaks, this means that if a CVSS with a score of 7 or above, then the build will fail. It also has some template code in the pom file, so that if you wanted to enforce 'x' test coverage stats, it's very easy to do so.